cleanup: inline journal tab into Book, delete standalone page
Journal.tsx had no direct route (already redirected to /book). Moved
its logic into Book.tsx as a local JournalTab component. Also fixed
Onboarding to navigate to /today instead of stale /workspace route.
Updated WEB_UI_GUIDE to remove Journal as a standalone page entry.
